- Refactor to more accessible format
