+++ messages/de.json
@@ -109,7 +109,7 @@
   },
 
   "theme": {
-    "light_switch": "Lichtschalter",
+    "light_switch": "Thema ändern",
     "light": "Licht",
     "dark": "dunkel",
     "system": "System"
