+++ messages/en.json
@@ -134,7 +134,7 @@
   },
 
   "theme": {
-    "light_switch": "Light switch",
+    "light_switch": "Change theme",
     "light": "Light",
     "dark": "Dark",
     "system": "System"
