+++ messages/hi.json
@@ -134,7 +134,7 @@
   },
 
   "theme": {
-    "light_switch": "प्रकाश स्विच",
+    "light_switch": "विषय को परिवर्तित करें",
     "light": "प्रकाश",
     "dark": "अँधेरा",
     "system": "सिस्टम"
