+++ src/lib/components/global/footer.svelte
